Mostly adds value estimate plus some fixes.
models.py adds value_estimate to the value heads
model_serialization.py, checkpoint_manager.py, tf_model_saver.py exports action, valueEstimate and action+valueEstimate to three separate model files
torch.py fixes so I can run learn.py from the python command line. probably don't need this if using the -m flag

Made some stuff public so I can access it from my scripts
Fix bug with Academy where multiple AcademySteppers could be in scene which breaks training.
Added some integrity checks to Academy
Fixed some bugs where CollectObservations is called outside the context of a step (between episodes).

+++ ml-agents/mlagents/trainers/model_saver/tf_model_saver.py
@@ -157,9 +157,23 @@ class TFModelSaver(BaseModelSaver):
         final_model_name = os.path.splitext(source_nn_path)[0]
 
         if SerializationSettings.convert_to_barracuda:
+            source_path = f"{final_model_name}_a.nn"
+            if os.path.isfile(source_path):
+                destination_path = f"{self.model_path}_a.nn"
+                shutil.copyfile(source_path, destination_path)
+            
+            source_path = f"{final_model_name}_v.nn"
+            if os.path.isfile(source_path):
+                source_path = f"{final_model_name}_v.nn"
+                destination_path = f"{self.model_path}_v.nn"
+                shutil.copyfile(source_path, destination_path)
+            
             source_path = f"{final_model_name}.nn"
-            destination_path = f"{self.model_path}.nn"
-            shutil.copyfile(source_path, destination_path)
+            if os.path.isfile(source_path):
+                source_path = f"{final_model_name}.nn"
+                destination_path = f"{self.model_path}.nn"
+                shutil.copyfile(source_path, destination_path)
+            
             logger.info(f"Copied {source_path} to {destination_path}.")
 
         if SerializationSettings.convert_to_onnx:
